Update structure parametres.c

+++ sources/structure parametres.c
@@ -1,11 +1,11 @@
 struct parametres{
-  int xi;
-  int yi;
-  int zi;
-  int dt;
-  int Tmax;
-  int σ;
-  int ρ;
-  int β;
+  float xi;
+  float yi;
+  float zi;
+  float dt;
+  float Tmax;
+  float σ;
+  float ρ;
+  float β;
 }
